refactor: extract language data, map_path, and instance helpers from config.py
- config_language_data.py: _LANGUAGE_TAGS dict, SUPPORTED_LANGUAGES list, _get_language_tags (~151 LOC)
- config_instances.py: get_sonarr_instances, get_radarr_instances, is_standalone_mode, get_media_server_instances (~144 LOC)
- config_utils.py: map_path function with circular-import-safe local import (~46 LOC)
- config.py: re-exports all moved symbols for full backwards compatibility (805 LOC, down from 1101)
- Add smoke tests confirming all public APIs survive the split
